requirements.txt
aiogram
psycopg2-binary
pyTelegramBotAPI